+++ CHANGELOG.md
@@ -3,7 +3,14 @@ Changes to Calva.
 
 ## [Unreleased]
 
-## [2.0.45] - 08.10.2019
+## [2.0.47] - 10.10.2019
+- [Support Jack-in without file open for single-rooted workspace](https://github.com/BetterThanTomorrow/calva/issues/366)
+- [Show argument list of fn](https://github.com/BetterThanTomorrow/calva/issues/238)
+- [Make code more robust in case Jack-in task fails](https://github.com/BetterThanTomorrow/calva/issues/367)
+- [Fix dimming out of stacked ignored forms](https://github.com/BetterThanTomorrow/calva/issues/385)
+- [The extension should specify the default schemes for document selectors](https://github.com/BetterThanTomorrow/calva/issues/368)
+
+## [2.0.46] - 08.10.2019
 - [Connect warnings and errors as popups](https://github.com/BetterThanTomorrow/calva/issues/356)
 - [Don't remove default indents when Calva is not the auto-formatter](https://github.com/BetterThanTomorrow/calva/pull/383)